Some directory and code changes
