+++ src/components/controls/PathChooser.svelte
@@ -4,7 +4,7 @@
   export let defaultPath = "";
   export let onEnd = () => {};
   const { ipcRenderer } = require("electron");
-  import { settings } from "../store";
+  import { settings } from "../../store";
   async function openDialog() {
     folderPath = await ipcRenderer.invoke("show-dialog", {
       dialogType: "openDirectory",
