+++ src/components/controls/SeekBar.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-  import { songsPlayer } from "../store";
+  import { songsPlayer } from "../../store";
   export let fullSize = 0;
   export let currentSize = 0;
   export let height = "7px";
@@ -7,19 +7,13 @@
   export let isSeekBar = false;
   function onClick(event: any) {
     if (event.target.firstChild) {
-      currentSize =
-        ((event.x - event.target.firstChild.offsetLeft) /
-          event.target.clientWidth) *
-        fullSize;
+      currentSize = ((event.x - event.target.firstChild.offsetLeft) / event.target.clientWidth) * fullSize;
       if (isSeekBar) {
         $songsPlayer.sound.seek(currentSize);
       }
       return;
     }
-    currentSize =
-      ((event.x - event.target.offsetLeft) /
-        event.target.parentNode.clientWidth) *
-      fullSize;
+    currentSize = ((event.x - event.target.offsetLeft) / event.target.parentNode.clientWidth) * fullSize;
     if (isSeekBar) {
       $songsPlayer.sound.seek(currentSize);
     }
