+++ src/components/controls/Toggle.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   export let checked = false;
   export let clickEvent = () => {};
-  import { currentTheme } from "../store";
+  import { currentTheme } from "../../store";
 </script>
 
 <main class:checked on:click={clickEvent} />
