+++ src/components/items/AlbumItem.svelte
@@ -1,0 +1,16 @@
+<script lang="ts">
+  export let album = {};
+  export let key;
+  import { UnknownAlbum } from "rhyme-icons";
+  import { push } from "svelte-spa-router";
+  import Item from "./Item.svelte";
+  const onClick = () => {
+    push(`/album-view?id=${key}`);
+  };
+</script>
+
+{#if album["songs"].length === 1}
+  <Item firstTitle={album["name"]} secondTitle={album["songs"].length + " Song" + (album["songs"].length > 1 ? "s" : "")} image={album["songs"][0].imgSrc} />
+{:else}
+  <Item firstTitle={album["name"]} secondTitle={album["songs"].length + " Song" + (album["songs"].length > 1 ? "s" : "")} component={UnknownAlbum} />
+{/if}
