+++ src/views/Settings.svelte
@@ -1,17 +1,9 @@
 <script lang="ts">
-  import PathChooser from "../controls/PathChooser.svelte";
-  import PopUpDialog from "../controls/PopUpDialog.svelte";
-  import Toggle from "../controls/Toggle.svelte";
+  import PathChooser from "../components/controls/PathChooser.svelte";
+  import PopUpDialog from "../components/controls/PopUpDialog.svelte";
+  import Toggle from "../components/controls/Toggle.svelte";
   import { setWatcher } from "../lib/RhymeUtils";
-  import {
-    settings,
-    songs,
-    currentSong,
-    recentlyPlayed,
-    songsPlayer,
-    themeManager,
-    currentTheme,
-  } from "../store";
+  import { settings, songs, currentSong, recentlyPlayed, songsPlayer, themeManager, currentTheme } from "../store";
 
   let folderPath = $settings["musicPath"];
   function changeMusicDir() {
@@ -55,12 +47,7 @@
   <ul>
     <li>
       <span> Music folder path </span>
-      <PathChooser
-        bind:folderPath
-        defaultPath={$settings["musicPath"]}
-        title="Choose the folder containing your music"
-        onEnd={changeMusicDir}
-      />
+      <PathChooser bind:folderPath defaultPath={$settings["musicPath"]} title="Choose the folder containing your music" onEnd={changeMusicDir} />
     </li>
     <li>
       <span>
@@ -71,10 +58,7 @@
     </li>
     <li>
       <span>Show notifications</span>
-      <Toggle
-        bind:checked={showNotifications}
-        clickEvent={toggleShowNotifications}
-      />
+      <Toggle bind:checked={showNotifications} clickEvent={toggleShowNotifications} />
     </li>
   </ul>
   <br />
