+++ src/widgets/BottomBar.svelte
@@ -1,25 +1,7 @@
 <script lang="ts">
-  import {
-    songsPlayer,
-    inAlbum,
-    currentTheme,
-    currentSong,
-    songPlaying,
-    shuffle,
-    repeat,
-    volume,
-  } from "../store";
-  import {
-    PauseCircleFilled,
-    PlayCircleFilled,
-    SkipPrevious,
-    SkipNext,
-    Repeat,
-    Shuffle,
-    VolumeOff,
-    VolumeUp,
-  } from "rhyme-icons";
-  import SeekBar from "../controls/SeekBar.svelte";
+  import { songsPlayer, inAlbum, currentTheme, currentSong, songPlaying, shuffle, repeat, volume } from "../store";
+  import { PauseCircleFilled, PlayCircleFilled, SkipPrevious, SkipNext, Repeat, Shuffle, VolumeOff, VolumeUp } from "rhyme-icons";
+  import SeekBar from "../components/controls/SeekBar.svelte";
 
   let currentTime = "";
   let duration = "";
@@ -88,11 +70,7 @@
 {#if $currentSong}
   <main>
     <div class="song-info">
-      <img
-        src={$currentSong["imgSrc"]}
-        style="display:{$currentSong['imgSrc'] ? 'block' : 'none'}"
-        alt=""
-      />
+      <img src={$currentSong["imgSrc"]} style="display:{$currentSong['imgSrc'] ? 'block' : 'none'}" alt="" />
       <div class="titles">
         <span><p class="ellipsis-text">{$currentSong["song"]}</p></span>
         <p class="ellipsis-text">
@@ -109,11 +87,7 @@
               shuffle.set(!$shuffle);
             }}
           >
-            <Shuffle
-              fill={$shuffle
-                ? $currentTheme["accentColor"]
-                : $currentTheme["textColor"]}
-            />
+            <Shuffle fill={$shuffle ? $currentTheme["accentColor"] : $currentTheme["textColor"]} />
           </div>
           <div
             on:click={() => {
@@ -137,10 +111,7 @@
                   $songsPlayer.pause();
                 }}
               >
-                <PauseCircleFilled
-                  size="36"
-                  fill={$currentTheme["textColor"]}
-                />
+                <PauseCircleFilled size="36" fill={$currentTheme["textColor"]} />
               </div>
             {/if}
           </div>
@@ -156,11 +127,7 @@
               repeat.set(!$repeat);
             }}
           >
-            <Repeat
-              fill={$repeat
-                ? $currentTheme["accentColor"]
-                : $currentTheme["textColor"]}
-            />
+            <Repeat fill={$repeat ? $currentTheme["accentColor"] : $currentTheme["textColor"]} />
           </div>
         </div>
         <div class="volume">
@@ -181,28 +148,15 @@
               <VolumeUp fill={$currentTheme["textColor"]} />
             </div>
           {/if}
-          <SeekBar
-            width="100px"
-            height="5px"
-            bind:currentSize={$volume}
-            fullSize={100}
-          />
+          <SeekBar width="100px" height="5px" bind:currentSize={$volume} fullSize={100} />
           {$volume}
         </div>
       </div>
       <div class="seekbar">
-        <p
-          style="color:{$currentTheme[
-            'accentColor'
-          ]};padding:0;margin:0;min-width:max-content;"
-        >
+        <p style="color:{$currentTheme['accentColor']};padding:0;margin:0;min-width:max-content;">
           {currentTime}
         </p>
-        <SeekBar
-          bind:fullSize={barSize["size"]}
-          bind:currentSize={barSize["current"]}
-          isSeekBar={true}
-        />
+        <SeekBar bind:fullSize={barSize["size"]} bind:currentSize={barSize["current"]} isSeekBar={true} />
         {duration}
       </div>
     </div>
