+++ src/widgets/Sidebar.svelte
